requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
django-allauth==0.36.0
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-discover-runner==1.0
django-livereload-server==0.3.2
django-rest-auth==0.9.3
django-role-permissions==2.2.0
djangorestframework==3.8.2
idna==2.7
oauthlib==2.1.0
python3-openid==3.1.0
pytz==2018.4
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
tornado==5.0.2
urllib3==1.23
